added abstract modifier to product class, made a method to showcase how many % you get discount of instead of showing decimals. cleaned up some things.
Created CurrencyHandler class to move the enum and  methods from the abstract Product class.

Made a potato mistake in my Products class not adding arguments to the constructor..  fixed it.
I had card coded the product list like a fucking idiot. spent an hour not understanding why it was working before and not now.

changed ProductInfo method from virtual to abstract.

added product and fixed the method to showcase and select items so there is no need to manually add items in the method every time a new item gets added or deleted.